refactor(cli): move cdktf json handling into @cdktf/commons
Closes #1814
