+++ packages/@cdktf/commons/src/config.ts
@@ -5,10 +5,11 @@
 
 import * as fs from "fs-extra";
 import * as path from "path";
-import { logger } from "./logging";
-import { isRegistryModule } from "./terraform-module";
 import { env } from "process";
 import { toPascalCase } from "codemaker";
+import { logger } from "./logging";
+import { isRegistryModule } from "./terraform-module";
+import { Errors } from "./errors";
 
 export enum Language {
   TYPESCRIPT = "typescript",
@@ -311,4 +312,93 @@ export function logTimespan(message: string) {
     const duration = end - start;
     logger.debug(`${message} took ${duration}ms`);
   };
+}
+
+function findFileAboveCwd(
+  file: string,
+  rootPath = process.cwd()
+): string | null {
+  const fullPath = path.resolve(rootPath, file);
+  if (fs.existsSync(fullPath)) {
+    return fullPath;
+  }
+
+  const parentDir = path.resolve(rootPath, "..");
+  if (fs.existsSync(parentDir) && parentDir !== rootPath) {
+    return findFileAboveCwd(file, parentDir);
+  }
+
+  return null;
+}
+export class CdktfConfig {
+  constructor(private cdktfConfigPath: string) {}
+
+  private readCdktfConfig(): Record<string, unknown> {
+    const cdktfConfig = require(this.cdktfConfigPath);
+    if (typeof cdktfConfig !== "object" || cdktfConfig === null) {
+      throw Errors.External(
+        "cdktf.json is malformed. The root must be a JSON object."
+      ); // TODO: define a schema and validate against it
+    }
+    return cdktfConfig;
+  }
+
+  private writeCdktfConfig(cdktfConfig: Record<string, unknown>) {
+    const cdktfConfigString = JSON.stringify(cdktfConfig, null, 2);
+    fs.writeFileSync(this.cdktfConfigPath, cdktfConfigString);
+  }
+
+  private getProperty(property: string): unknown {
+    const cdktfConfig = this.readCdktfConfig();
+    return cdktfConfig[property] || (CONFIG_DEFAULTS as any)[property];
+  }
+
+  public get language(): Language {
+    const rawLanguage = this.getProperty("language");
+    const language = Object.values(Language).find(
+      (value) => value === rawLanguage
+    );
+    if (!language)
+      throw Errors.Usage(
+        `${rawLanguage} is not a valid language. It must be one of ${Object.values(
+          Language
+        )}`
+      );
+
+    return language;
+  }
+
+  public get codeMakerOutput(): string {
+    return this.getProperty("codeMakerOutput") as string;
+  }
+
+  public get terraformProviders(): (TerraformDependencyConstraint | string)[] {
+    const providers = this.getProperty("terraformProviders");
+    if (!Array.isArray(providers)) return [];
+    return providers;
+  }
+
+  public writeTerraformProviders(
+    providers: (TerraformDependencyConstraint | string)[]
+  ) {
+    const cdktfConfig = this.readCdktfConfig();
+    cdktfConfig.terraformProviders = providers;
+    this.writeCdktfConfig(cdktfConfig);
+  }
+
+  public get projectDirectory(): string {
+    return path.dirname(this.cdktfConfigPath);
+  }
+
+  public static read(path: string = process.cwd()): CdktfConfig {
+    const cdktfConfigPath = findFileAboveCwd("cdktf.json", path);
+    if (!cdktfConfigPath) {
+      throw Errors.External(
+        "Could not find cdktf.json. Make sure there is a cdktf.json file in the current directory or one of its parents."
+      );
+    }
+    logger.trace(`cdktf.json found at ${cdktfConfigPath}`);
+
+    return new CdktfConfig(cdktfConfigPath);
+  }
 }
